feat(api): add npipe support for local endpoint
