+++ api/http/internal/socket.go
@@ -1,0 +1,51 @@
+// +build !windows
+
+package internal
+
+import (
+	"io"
+	"net"
+	"net/http"
+	"net/http/httputil"
+)
+
+// NewFileProxy returns a HTTP handler writing to a Unix socket.
+func NewFileProxy(path string) http.Handler {
+	return &unixSocketHandler{path}
+}
+
+// NewUnixSocketHandler returns a HTTP handler writing to a Unix socket.
+func NewUnixSocketHandler(path string) http.Handler {
+	return &unixSocketHandler{path}
+}
+
+// unixSocketHandler represents a handler to proxy HTTP requests via a unix:// socket
+type unixSocketHandler struct {
+	path string
+}
+
+func (h *unixSocketHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	conn, err := net.Dial("unix", h.path)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	c := httputil.NewClientConn(conn, nil)
+	defer c.Close()
+
+	res, err := c.Do(r)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	defer res.Body.Close()
+
+	for k, vv := range res.Header {
+		for _, v := range vv {
+			w.Header().Add(k, v)
+		}
+	}
+	if _, err := io.Copy(w, res.Body); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+}
